requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.0
openai==1.12.0
numpy==1.26.3
pandas==2.1.4
tqdm==4.66.1